--- projeto-frontend-senai-develop/src/app/shared/interfaces/produto.interface.ts
@@ -1,9 +1,0 @@
-export interface Produto {
-    id: string,
-    urlImg: string,
-    valor: number,
-    titulo: string,
-    desconto?: number,
-    vendidos: number,
-    descricao: string
-  }
