+++ src/app/app.component.html
@@ -1,5 +1,5 @@
 <div class="container">
-  <app-header></app-header>
+
   <router-outlet id="router"/>
   <app-footer></app-footer>
   <!-- <app-card-produto [urlImg]="'../assets/xbox.jpg'" [valor]="4000" [titulo]="'Xbox Série S - 1TB (preto)'"/> -->
