+++ src/app/busca-produtos/busca-produtos.component.html
@@ -1,4 +1,5 @@
-<!-- <app-header (buttonClick)="buttonClicked($event)"></app-header> -->
+
+<app-header (buttomClick)="buttonClicked($event)"></app-header>
 
 <fieldset class="exibicao">
     <legend> Produtos:</legend>
