+++ src/app/busca-produtos/busca-produtos.component.scss
@@ -1,16 +1,8 @@
-.header {
-    height: 120px;
-    border: 1px solid black;
-    margin-left: 20px;
-    margin-right: 20px;
-    margin-bottom: 80px;
-    margin-top: 20px;
-}
-
 .exibicao {
     margin-left: 20px;
     margin-right: 20px;
     margin-top: 80px;
+    margin-bottom: 80px;
 }
 
 .alert {
