+++ src/app/detalhes-produto/detalhes-produto.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DetalhesProdutoComponent } from './detalhes-produto.component';
+
+describe('DetalhesProdutoComponent', () => {
+  let component: DetalhesProdutoComponent;
+  let fixture: ComponentFixture<DetalhesProdutoComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [DetalhesProdutoComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(DetalhesProdutoComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
