+++ src/app/home/home.component.html
@@ -1,3 +1,4 @@
+<app-header></app-header>
 <div class="box-oferta-dia">
     <div class="titulo-oferta">
         <p>Oferta do Dia</p>
