+++ src/app/shared/interfaces/produto.interface.ts
@@ -1,0 +1,9 @@
+export interface Produto {
+    id: string,
+    urlImg: string,
+    valor: number,
+    titulo: string,
+    desconto?: number,
+    vendidos: number,
+    descricao: string
+  }
